chore: some cleanup of old files, split HUDExtensionModule, debug logs
